Advanced Task/requirements.txt
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gymnasium==0.29.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sympy==1.12
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.10.0
